Fix some nrof issues when many items. Always return real money value.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@7633 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/c_object.c
@@ -1518,7 +1518,7 @@ void examine(object *op, object *tmp) {
 
     if(tmp->weight) {
 	snprintf(buf, sizeof(buf), tmp->nrof > 1 ? "They weigh %3.3f kg." : "It weighs %3.3f kg.",
-            tmp->weight*(tmp->nrof?tmp->nrof:1)/1000.0);
+            tmp->weight*((float)(tmp->nrof?tmp->nrof:1)/1000.0));
 	draw_ext_info(NDI_UNIQUE, 0,op,MSG_TYPE_COMMAND, MSG_TYPE_COMMAND_EXAMINE,
 		      buf, NULL);
     }
